+++ src/main/java/vn/hactanco/socialnetwork/service/PostService.java
@@ -4,7 +4,9 @@ import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Map;
 import java.util.UUID;
+import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
@@ -14,6 +16,8 @@ import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 
 import lombok.RequiredArgsConstructor;
+import vn.hactanco.socialnetwork.dto.PostMediaResponseDTO;
+import vn.hactanco.socialnetwork.dto.PostResponseDTO;
 import vn.hactanco.socialnetwork.enums.MediaType;
 import vn.hactanco.socialnetwork.exception.ResourceNotFoundException;
 import vn.hactanco.socialnetwork.model.Post;
@@ -130,12 +134,45 @@ public class PostService {
 		postRepository.save(post);
 	}
 
-	public Page<Post> getFeed(int page, int size) {
+	private PostResponseDTO convertToDTO(Post post) {
+//		private Long postId;
+//		private String content;
+//		private Instant createdAt;
+//	
+//		private Long userId;
+//		private String userName;
+//		private String userAvatar;
+//		private List<PostMediaResponseDTO> medias;
+
+		List<PostMediaResponseDTO> mediaDTOS = post.getMedias().stream().map(media -> PostMediaResponseDTO.builder()
+				.mediaUrl(media.getMediaUrl()).mediaType(media.getMediaType()).build()).toList();
+
+		return PostResponseDTO.builder().postId(post.getId()).content(post.getContent()).createdAt(post.getCreatedAt())
+				.userId(post.getUser().getId()).userName(post.getUser().getName())
+				.userAvatar(post.getUser().getAvatar()).medias(mediaDTOS).build();
+	}
+
+	public List<PostResponseDTO> getFeedDTO(int page, int size) {
 
 		Pageable pageable = PageRequest.of(page, size);
 
-		return postRepository.findAllByOrderByCreatedAtDesc(pageable);
+		Page<Long> postIdsPage = postRepository.findPostIds(pageable);
+		List<Long> ids = postIdsPage.getContent();
+		List<Post> posts = postRepository.findByIdInWithUserAndMedia(ids);
+		// map lại theo id
+		Map<Long, Post> map = posts.stream().collect(Collectors.toMap(Post::getId, p -> p));
+		// giữ đúng thứ tự từ query 1 (DESC)
+		List<Post> sortedPosts = ids.stream().map(map::get).toList();
+		return sortedPosts.stream().map(this::convertToDTO).toList();
+
 	}
+
+//	public Page<Post> getFeed(int page, int size) {
+//
+//		Pageable pageable = PageRequest.of(page, size);
+//
+//		return postRepository.findAllByOrderByCreatedAtDesc(pageable);
+//	}
 
 	public void deletePost(Long postId, User user) {
 
@@ -181,4 +218,77 @@ public class PostService {
 
 		postRepository.save(post);
 	}
+
+	public void updatePost(Long postId, String content, MultipartFile[] files, User user) throws IOException {
+
+		Post post = postRepository.findById(postId)
+				.orElseThrow(() -> new ResourceNotFoundException("Post không tồn tại"));
+
+		if (!post.getUser().getId().equals(user.getId())) {
+			throw new RuntimeException("Không có quyền sửa bài viết");
+		}
+
+		// update content
+		if (content == null || content.trim().isEmpty()) {
+			throw new ResourceNotFoundException("Nội dung không được trống");
+		}
+
+		post.setContent(content.trim());
+
+		// 👉 nếu có upload file mới → xóa media cũ + thêm mới
+		if (files != null && files.length > 0 && !files[0].isEmpty()) {
+
+			// XÓA FILE CŨ
+			if (post.getMedias() != null) {
+				for (PostMedia media : post.getMedias()) {
+					String filePath = uploadDir + media.getMediaUrl().replace("/uploads/", "");
+					new File(filePath).delete();
+				}
+			}
+
+			List<PostMedia> newMedias = new ArrayList<>();
+
+			for (MultipartFile file : files) {
+
+				if (file.isEmpty())
+					continue;
+
+				String originalName = file.getOriginalFilename();
+				String extension = originalName.substring(originalName.lastIndexOf(".") + 1).toLowerCase();
+
+				List<String> imageExt = List.of("jpg", "jpeg", "png", "gif", "webp");
+				List<String> videoExt = List.of("mp4", "mov", "avi", "webm");
+
+				boolean isImage = imageExt.contains(extension);
+				boolean isVideo = videoExt.contains(extension);
+
+				String fileName = UUID.randomUUID() + "_" + originalName;
+
+				File dest;
+				String mediaUrl;
+
+				if (isImage) {
+					dest = new File(uploadDir + "images/", fileName);
+					mediaUrl = "/uploads/images/" + fileName;
+				} else {
+					dest = new File(uploadDir + "videos/", fileName);
+					mediaUrl = "/uploads/videos/" + fileName;
+				}
+
+				file.transferTo(dest);
+
+				PostMedia media = new PostMedia();
+				media.setMediaUrl(mediaUrl);
+				media.setMediaType(isImage ? MediaType.IMAGE : MediaType.VIDEO);
+				media.setPost(post);
+
+				newMedias.add(media);
+			}
+
+			post.setMedias(newMedias);
+		}
+
+		postRepository.save(post);
+	}
+
 }